pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autocontrol-scientist"
version = "1.0.0"
description = "Multi-agent control systems research automation platform"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "AutoControl-Scientist Team" }]
keywords = ["control-systems", "automation", "llm", "agents", "research"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "PyQt6>=6.4.0",
    "cryptography>=41.0.0",
    "aiohttp>=3.10.0",
    "PyYAML>=6.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-qt>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML",
    "types-aiofiles",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
]
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]
matlab = []

[project.scripts]
autocontrol = "main:main"
autocontrol-api = "api_main:app"

[project.urls]
Homepage = "https://github.com/yourusername/autocontrol-scientist"
Documentation = "https://github.com/yourusername/autocontrol-scientist#readme"
Repository = "https://github.com/yourusername/autocontrol-scientist"
Issues = "https://github.com/yourusername/autocontrol-scientist/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "output*", "__pycache__*", ".git*"]

[tool.ruff]
target-version = "py39"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    "*.pyc",
    ".venv",
    "venv",
    "output",
    ".pytest_cache",
    ".ruff_cache",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "RUF"]
ignore = ["E501", "N806"]

[tool.ruff.lint.isort]
known-first-party = ["autocontrol_scientist"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true
pretty = true
exclude = "(?x)(^output/|^__pycache__/|\\.pyc$)"

[[tool.mypy.overrides]]
module = [
    "PyQt6.*",
    "aiohttp.*",
    "openai.*",
    "anthropic.*",
    "scholarly.*",
    "matlab.*",
    "matlab.engine.*",
    "fastapi.*",
    "uvicorn.*",
    "pydantic.*",
    "opentelemetry.*",
    "matplotlib.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-fail-under=60",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["api", "core", "config_manager.py"]
omit = [
    "tests/*",
    "*/__pycache__/*",
    "main.py",
]

[tool.coverage.report]
fail_under = 60
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
